code cleanup

+++ src/include/capa.h
@@ -4,10 +4,10 @@
 
 typedef struct capa_s {
     uint32_t* imagen;
-    uint16_t ancho;
-    uint16_t altura;
-    uint16_t x;
-    uint16_t y;
+    uint32_t ancho;
+    uint32_t altura;
+    uint32_t x;
+    uint32_t y;
 } capa_t;
 
 capa_t combinar_capas(capa_t capa_superior, capa_t capa_inferior);
